# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.JSONMap:
    additionalProperties: true
    type: object
  models.Payment:
    properties:
      amount_in_paise:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      gateway:
        $ref: '#/definitions/models.PaymentGateway'
      gateway_response:
        $ref: '#/definitions/models.JSONMap'
      id:
        type: string
      notes:
        type: string
      paid_at:
        type: string
      plan:
        $ref: '#/definitions/models.Plan'
      plan_id:
        type: string
      razorpay_order_id:
        type: string
      razorpay_payment_id:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.PaymentGateway:
    enum:
    - RAZORPAY
    - MANUAL
    type: string
    x-enum-varnames:
    - GatewayRazorpay
    - GatewayManual
  models.PaymentStatus:
    enum:
    - SUCCESS
    - FAILED
    - REFUNDED
    - PENDING
    type: string
    x-enum-varnames:
    - PaymentSuccess
    - PaymentFailed
    - PaymentRefunded
    - PaymentPending
  models.Photo:
    properties:
      category:
        type: string
      file_key:
        type: string
      id:
        type: string
      is_published:
        type: boolean
      mime_type:
        type: string
      size_bytes:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      uploaded_at:
        type: string
      url:
        description: URL is computed at request time from FileKey — not stored in
          DB.
        type: string
      video_id:
        type: string
    type: object
...
  models.Plan:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_days:
        type: integer
      has_e4:
        type: boolean
      has_mecad:
        type: boolean
      has_willcom:
        type: boolean
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price_in_paise:
        type: integer
      updated_at:
        type: string
    type: object
  models.PlaybackLog:
    properties:
      completed:
        type: boolean
      id:
        type: string
      played_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      video:
        $ref: '#/definitions/models.Video'
      video_id:
        type: string
      watched_seconds:
        type: integer
      watermark_id:
        type: string
    type: object
  models.Subscription:
    properties:
      activated_by:
        type: string
      created_at:
        type: string
      expiry_date:
        type: string
      expiry_warning_sent:
        type: boolean
      id:
        type: string
      plan:
        $ref: '#/definitions/models.Plan'
      plan_id:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.SubscriptionStatus'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      joined_at:
        type: string
      name:
        type: string
      payments:
        items:
          $ref: '#/definitions/models.Payment'
        type: array
      phone:
        type: string
      playback_logs:
        items:
          $ref: '#/definitions/models.PlaybackLog'
        type: array
      sessions:
        items:
          $ref: '#/definitions/models.UserSession'
        type: array
      status:
        $ref: '#/definitions/models.UserStatus'
      subscriptions:
        items:
          $ref: '#/definitions/models.Subscription'
        type: array
      updated_at:
        type: string
    type: object
  models.UserSession:
    properties:
      created_at:
        type: string
      device_info:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_active_at:
        type: string
      location:
        type: string
      revoked_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.UserStatus:
    enum:
    - ACTIVE
    - SUSPENDED
    - EXPIRED
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusSuspended
    - UserStatusExpired
  models.Video:
    properties:
      category:
        $ref: '#/definitions/models.VideoCategory'
      description:
        type: string
      duration_seconds:
        type: integer
      file_key:
        type: string
      hls_key:
        type: string
      id:
        type: string
      is_free:
        type: boolean
      is_intro:
        type: boolean
      is_preview:
        type: boolean
      lqip:
        type: string
      mp4_key_240p:
        type: string
      mp4_key_360p:
        type: string
      mp4_key_480p:
        type: string
      mp4_key_720p:
        type: string
      mp4_key_1080p:
        type: string
      mp4_size_240p:
        type: integer
      mp4_size_360p:
        type: integer
      mp4_size_480p:
        type: integer
      mp4_size_720p:
        type: integer
      mp4_size_1080p:
        type: integer
      playback_logs:
        items:
          $ref: '#/definitions/models.PlaybackLog'
        type: array
      preview_url:
        type: string
      status:
        $ref: '#/definitions/models.VideoStatus'
      thumbnail_url:
        type: string
      title:
        type: string
      transcode_error:
        type: string
      updated_at:
        type: string
      uploaded_at:
        type: string
    type: object
  models.VideoCategory:
    enum:
    - WILLCOM
    - E4
    - MECAD
    type: string
    x-enum-varnames:
    - CategoryWillcom
    - CategoryE4
    - CategoryMecad
...
  models.VideoProgress:
    properties:
      position_seconds:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
      video_id:
        type: string
    type: object
  models.VideoStatus:
    enum:
    - DRAFT
    - PROCESSING
    - PUBLISHED
    - ERROR
    type: string
    x-enum-varnames:
    - VideoStatusDraft
    - VideoStatusProcessing
    - VideoStatusPublished
    - VideoStatusError
...
  user_videos.videoWithAccess:
    properties:
      accessible:
        type: boolean
      category:
        $ref: '#/definitions/models.VideoCategory'
      description:
        type: string
      duration_seconds:
        type: integer
      file_key:
        type: string
      hls_key:
        type: string
      id:
        type: string
      is_free:
        type: boolean
      is_intro:
        type: boolean
      is_preview:
        type: boolean
      lqip:
        type: string
      mp4_key_240p:
        type: string
      mp4_key_360p:
        type: string
      mp4_key_480p:
        type: string
      mp4_key_720p:
        type: string
      mp4_key_1080p:
        type: string
      mp4_size_240p:
        type: integer
      mp4_size_360p:
        type: integer
      mp4_size_480p:
        type: integer
      mp4_size_720p:
        type: integer
      mp4_size_1080p:
        type: integer
      playback_logs:
        items:
          $ref: '#/definitions/models.PlaybackLog'
        type: array
      preview_url:
        type: string
      progress_seconds:
        type: integer
      status:
        $ref: '#/definitions/models.VideoStatus'
      thumbnail_url:
        type: string
      title:
        type: string
      transcode_error:
        type: string
      updated_at:
        type: string
      uploaded_at:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: MarketKit
  description: REST API for the MarketKit marketplace platform. Admin routes require
    a Bearer JWT token. User routes require a separate User Bearer JWT token.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: MarketKit API
  version: "1.0"
paths:
...
  /user/videos:
    get:
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: Category filter (WILLCOM, E4, MECAD)
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user_videos.videoWithAccess'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: List published videos with access flags
      tags:
      - User Videos
...
  /user/videos/{id}/download-url:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Quality: 240p | 360p | 480p | 720p | 1080p (default 720p)'
        in: query
        name: quality
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Get a signed download URL for a specific quality MP4
      tags:
      - User Videos
  /user/videos/{id}/photos:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Photo'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: List photos attached to a video
      tags:
      - User Videos
  /user/videos/{id}/playback-log:
    post:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Log a video playback event
      tags:
      - User Videos
  /user/videos/{id}/progress:
    post:
      consumes:
      - application/json
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: position_seconds
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VideoProgress'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Save video playback progress
      tags:
      - User Videos
  /user/videos/{id}/qualities:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: List available download qualities with real file sizes
      tags:
      - User Videos
...
  /user/videos/{id}/stream:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "307":
          description: Redirect to signed video URL
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Stream a video (redirect to signed URL)
      tags:
      - User Videos
  /user/videos/{id}/stream-url:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Get signed video stream URL as JSON
      tags:
      - User Videos
...
securityDefinitions:
  AdminAuth:
    description: Admin JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    description: User JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"